refactor: DTO 패키지 재분배
