+++ src/main/java/wooteco/subway/admin/controller/advice/SubwayControllerAdvice.java
@@ -4,7 +4,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import wooteco.subway.admin.dto.ExceptionResponse;
+import wooteco.subway.admin.dto.response.ExceptionResponse;
 import wooteco.subway.admin.exception.*;
 
 @RestControllerAdvice
