+++ src/main/java/wooteco/subway/admin/dto/request/LineStationCreateRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.dto;
+package wooteco.subway.admin.dto.request;
 
 public class LineStationCreateRequest {
     private Long preStationId;
