+++ src/main/java/wooteco/subway/admin/dto/response/ExceptionResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.dto;
+package wooteco.subway.admin.dto.response;
 
 public class ExceptionResponse {
 	private String message;
