+++ src/main/java/wooteco/subway/admin/dto/response/WholeSubwayResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.dto;
+package wooteco.subway.admin.dto.response;
 
 import java.util.List;
