+++ src/test/java/wooteco/subway/admin/acceptance/LineAcceptanceTest.java
@@ -2,8 +2,8 @@ package wooteco.subway.admin.acceptance;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import wooteco.subway.admin.dto.LineDetailResponse;
-import wooteco.subway.admin.dto.LineResponse;
+import wooteco.subway.admin.dto.response.LineDetailResponse;
+import wooteco.subway.admin.dto.response.LineResponse;
 
 import java.time.LocalTime;
 import java.util.List;
